changes to featureset. FeatureSet is refactored to FeatureExtractor, which now does not contain the parent set. All it contains is the indices included in the feature set. This way, FeatureExtractor can apply the same thing to ANY parent set.
Because of this, modelversion has also been changed. Instead of a single feature set, it now takes a parent set AND a feature extractor. To score the model, the features and labels arrays are obtained by using the FeatureExtractor methods on the parent set. The same happens when training the model. To predict, any arbitrary parent set (that is of the same format as the original training parent set) can be used and FeatureExtractor will automatically be applied to it.

Slight changes to model collection is now required because it has to pass a parent set and feature extractor to create more modelversion instances, as opposed to sending a single feature set.